+++ src/resources/views/zone/show.blade.php
@@ -8,18 +8,7 @@
 
     <div class="card mb-4">
         <div class="card-body">
-            @foreach($zone->members as $member)
-                <span @class(['badge badge-pill', 'badge-info' => $member->isCountry(), 'badge-success' => $member->isProvince()])>
-                    {{ $member->member->name }}
-                </span>
-            @endforeach
-
-            @can('create zones')
-                <a href="{{ route('vanilo.admin.zone.create') }}"
-                   class="btn btn-success btn-sm mb-1"
-                   title="{{ __('Add member') }}">{!! icon('+') !!}</a>
-            @endcan
-
+            @include('vanilo::zone-member._index', ['zoneMembers' => $zone->members])
         </div>
     </div>
 
@@ -29,7 +18,7 @@
                 <a href="{{ route('vanilo.admin.zone.edit', $zone) }}" class="btn btn-outline-primary">{{ __('Edit Zone') }}</a>
             @endcan
 
-            @can('delete carriers')
+            @can('delete zones')
                 {!! Form::open([
                         'route' => ['vanilo.admin.zone.destroy', $zone],
                         'method' => 'DELETE',